Fix warnings.

+++ physics/n_body_system_body.hpp
@@ -187,8 +187,8 @@ void NBodySystem<InertialFrame>::ComputeGravitationalAccelerations(
     std::vector<Length> const& q,
     std::vector<Acceleration>* result) {
   result->assign(result->size(), Acceleration());
-  int const number_of_massive_trajectories = massive_trajectories.size();
-  int const number_of_massless_trajectories = massless_trajectories.size();
+  size_t const number_of_massive_trajectories = massive_trajectories.size();
+  size_t const number_of_massless_trajectories = massless_trajectories.size();
 
   for (std::size_t b1 = 0, three_b1 = 0;
        b1 < number_of_massive_trajectories;
@@ -220,7 +220,7 @@ void NBodySystem<InertialFrame>::ComputeGravitationalAccelerations(
       (*result)[three_b2 + 1] += Δq1 * μ1OverRSquared;
       (*result)[three_b2 + 2] += Δq2 * μ1OverRSquared;
     }
-    for (int b2 = number_of_massive_trajectories;
+    for (size_t b2 = number_of_massive_trajectories;
          b2 < number_of_massive_trajectories + number_of_massless_trajectories;
          ++b2) {
       std::size_t const three_b2 = 3 * b2;
@@ -241,7 +241,7 @@ void NBodySystem<InertialFrame>::ComputeGravitationalAccelerations(
   }
 
   // Finally, take into account the intrinsic accelerations.
-  for (int b2 = number_of_massive_trajectories;
+  for (size_t b2 = number_of_massive_trajectories;
        b2 < number_of_massive_trajectories + number_of_massless_trajectories;
        ++b2) {
     std::size_t const three_b2 = 3 * b2;
